+++ lib/ductwork/models/job.rb
@@ -100,7 +100,7 @@ module Ductwork
         job_klass: klass
       )
       args = JSON.parse(input_args)["args"]
-      instance = Object.const_get(klass).build_for_execution(step, args)
+      instance = Object.const_get(klass).build_for_execution(step.pipeline_id, args)
       run = execution.create_run!(
         started_at: Time.current
       )