+++ lib/ductwork/models/step.rb
@@ -26,9 +26,9 @@ module Ductwork
          expand: "expand",
          collapse: "collapse"
 
-    def self.build_for_execution(record, *args, **kwargs)
+    def self.build_for_execution(pipeline_id, *args, **kwargs)
       instance = allocate
-      instance.instance_variable_set(:@pipeline_id, record.pipeline_id)
+      instance.instance_variable_set(:@pipeline_id, pipeline_id)
       instance.send(:initialize, *args, **kwargs)
       instance
     end